+++ relay-common-actors/src/lib.rs
@@ -1,0 +1,8 @@
+//! Common actors for Relay services
+#![warn(missing_docs)]
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/getsentry/relay/master/artwork/relay-icon.png",
+    html_favicon_url = "https://raw.githubusercontent.com/getsentry/relay/master/artwork/relay-icon.png"
+)]
+
+pub mod controller;